+++ include/sho_potential.hxx
@@ -4,6 +4,7 @@
 #include <vector> // std::vector<T>
 #include <fstream> // std::ifstream
 #include <algorithm> // std::max
+#include <cstdint> // int8_t
 
 #include "status.hxx" // status_t
 #include "inline_math.hxx" // set
@@ -309,7 +310,7 @@ namespace sho_potential {
       view2D<double> xyzZ;
       int natoms{0};
       double cell[3] = {0, 0, 0}; 
-      int bc[3] = {-7, -7, -7};
+      int8_t bc[3] = {-7, -7, -7};
       { // scope: read atomic positions
           stat += geometry_analysis::read_xyz_file(xyzZ, natoms, geo_file, cell, bc, 0);
           if (echo > 2) std::printf("# found %d atoms in file \"%s\" with cell=[%.3f %.3f %.3f] %s and bc=[%d %d %d]\n",
