+++ src/sho_hamiltonian.cxx
@@ -598,7 +598,7 @@ namespace sho_hamiltonian {
       view2D<double> xyzZ;
       int natoms{0}; // number of atoms
       double cell[3] = {0, 0, 0}; // rectangular cell
-      int bc[3] = {-7, -7, -7}; // boundary conditions
+      int8_t bc[3] = {-7, -7, -7}; // boundary conditions
       { // scope: read atomic positions
           stat += geometry_analysis::read_xyz_file(xyzZ, natoms, geo_file, cell, bc, 0);
           if (echo > 2) std::printf("# found %d atoms in file \"%s\" with cell=[%.3f %.3f %.3f] %s and bc=[%d %d %d]\n",
